Revert last commit
The re-calculation of the common mode, and therefore
pedestal and noise, at the postprocessing level (open_sesame)
will introduce extra processing time and some re-factorization
of the post-processing classes, which seems to be an unnecessary
approach. Masking channels (and automasking) at the fortythieves
level seems to be more reasonable. The user can always reprocess
with fortythieves without masking if needed (the processsing times
are acceptables so far).